requirements.txt
python-dotenv
tldextract
pathlib
pandas
ail-typo-squatting
dnspython
